requirements.txt
eth-ape==0.7.11
web3>=6,<7
python-dotenv>=1.0
